Prevent minimizing database multiple times when checksum matches

+++ pihole-backup.sh
@@ -21,6 +21,7 @@ if [ -f "${CONFIG_FILE}" ]; then
 fi
 
 command -v rclone >/dev/null 2>&1 || { echo "Please install Rclone!"; exit 1; }
+command -v md5sum >/dev/null 2>&1 || { echo "Please install md5sum!"; exit 1; }
 
 PID=$(cat ${LOCKFILE} 2> /dev/null || echo '')
 if [ -e ${LOCKFILE} ] && [ ! -z "$PID" ] && kill -0 $PID; then
@@ -39,9 +40,21 @@ trap onInterruptOrExit EXIT
 [ -f "$BACKUP_LIST_FILE" ] || { echo "Missing backup list file: $BACKUP_LIST_FILE"; exit 1; }
 
 if [ -f "${GRAVITYDB_FILE}" ]; then
-	echo "Minimizing gravity.db..."
-	cp -f ${GRAVITYDB_FILE} ${GRAVITYDB_MIN_FILE}
-	sqlite3 ${GRAVITYDB_MIN_FILE} "DELETE FROM gravity; VACUUM;"
+	PREVIOUS_CHECKSUM=
+	if [ -f "${GRAVITYDB_MIN_FILE}.cache" ]; then
+		PREVIOUS_CHECKSUM=`cat "${GRAVITYDB_MIN_FILE}.cache"`
+	fi
+
+	CURRENT_CHECKSUM=`md5sum ${GRAVITYDB_FILE} | awk '{ print $1 }'`
+
+	if [ "${CURRENT_CHECKSUM}" != "${PREVIOUS_CHECKSUM}" ]; then
+		echo "Minimizing gravity database file size..."
+		
+		cp -f ${GRAVITYDB_FILE} ${GRAVITYDB_MIN_FILE}
+		sqlite3 ${GRAVITYDB_MIN_FILE} "DELETE FROM gravity; VACUUM;"
+		
+		echo ${CURRENT_CHECKSUM} > "${GRAVITYDB_MIN_FILE}.cache"
+	fi
 fi
 
 echo "Backing up now..."
